WIP change timestamp to int64

+++ genesis/genesis.go
@@ -133,7 +133,7 @@ func validateConfig(networkID uint32, config *Config, stakingCfg *StakingConfig)
 		return errNoSupply
 	}
 
-	startTime := time.Unix(int64(config.StartTime), 0)
+	startTime := time.Unix(config.StartTime, 0)
 	if time.Since(startTime) < 0 {
 		return fmt.Errorf(
 			"%w: %s",
@@ -335,7 +335,7 @@ func FromConfig(config *Config) ([]byte, ids.ID, error) {
 		return nil, ids.ID{}, fmt.Errorf("couldn't generate AVAX asset ID: %w", err)
 	}
 
-	genesisTime := time.Unix(int64(config.StartTime), 0)
+	genesisTime := time.Unix(config.StartTime, 0)
 	initialSupply, err := config.InitialSupply()
 	if err != nil {
 		return nil, ids.ID{}, fmt.Errorf("couldn't calculate the initial supply: %w", err)
